requirements.txt
# requirements.txt

# --- Core Framework ---
Django>=4.2,<4.3  # LTS version, fully supported
psycopg2-binary>=2.9,<2.10  # Use the binary for easier installation
django-phonenumber-field[phonenumbers]>=7.3.0,<7.4 # For phone number validation
Babel>=2.12.1,<2.13 # Required for phonenumber_field widget


# --- Utilities ---
Pillow>=10.4,<10.5  # Critical security update
requests>=2.32,<2.33
openpyxl>=3.0,<3.2

twilio>=8.0,<9.0 # For sending WhatsApp/SMS verification
django-ratelimit>=4.0,<5.0 # For rate limiting requests
django-redis>=5.4.0
django-import-export>=3.3

django-tinymce>=3.6.1,<3.7
